mission : 시니어 미션 - 뷰 그리기
