+++ mission/thread/app/src/main/AndroidManifest.xml
@@ -23,6 +23,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".ThreadActivity" android:exported="false"/>
+        <activity android:name=".EditMemoActivity" android:exported="false"/>
     </application>
 
 </manifest>
