+++ mission/thread/app/src/main/java/com/example/myapplication/EditMemoActivity.kt
@@ -1,0 +1,16 @@
+package com.example.myapplication
+
+import android.app.Activity
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.example.myapplication.databinding.ActivityEditMemoBinding
+
+class EditMemoActivity : AppCompatActivity() {
+    lateinit var binding: ActivityEditMemoBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityEditMemoBinding.inflate(layoutInflater)
+
+    }
+}
