+++ mission/thread/app/src/main/res/drawable/background_button.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android" android:shape="rectangle">
+    <solid android:color="@color/sky_blue"/>
+    <corners android:radius="16dp"/>
+</shape>
