+++ mission/thread/app/src/main/res/drawable/background_memo.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android" android:shape="rectangle">
+    <stroke android:color="@color/black" android:width="2dp"/>
+    <corners android:radius="16dp"/>
+    <solid android:color="@color/gray"/>
+  
+</shape>
