+++ mission/thread/app/src/main/res/layout/activity_edit_memo.xml
@@ -1,30 +1,42 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:padding="16dp">
+
     <TextView
         android:id="@+id/tv_title"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Memo"
+        android:text="Edit Memo"
         android:textSize="28sp"
         android:textStyle="bold"
         android:textColor="@color/black"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintStart_toStartOf="parent"/>
-    <TextView
+
+    <androidx.appcompat.widget.AppCompatEditText
+        android:id="@+id/et_memo"
         android:layout_width="match_parent"
         android:layout_height="300dp"
-        android:background="@drawable/background_button"
-        android:backgroundTint="@color/gray"
+        android:background="@drawable/background_memo"
         android:textSize="16sp"
+        android:hint="메모를 입력하시오"
         android:gravity="top"
         android:padding="12dp"
         android:layout_marginTop="12dp"
         app:layout_constraintTop_toBottomOf="@id/tv_title"/>
-
+    
+    <androidx.appcompat.widget.AppCompatButton
+        android:id="@+id/btn_save"
+        android:layout_width="match_parent"
+        android:layout_height="60dp"
+        android:background="@drawable/background_button"
+        android:text="SAVE"
+        android:textSize="16sp"
+        app:layout_constraintBottom_toBottomOf="parent"/>
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
