+++ mission/thread/app/src/main/res/values/colors.xml
@@ -7,4 +7,6 @@
     <color name="teal_700">#FF018786</color>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
+    <color name="gray">#f1f3f5</color>
+    <color name="sky_blue">#b9e0fd</color>
 </resources>
